refactor(core): split node deps
