+++ packages/core/package.json
@@ -37,6 +37,7 @@
     "pathe": "^2.0.3",
     "postgres": "^3.4.7",
     "telegram": "^2.26.22",
+    "uuid": "^11.1.0",
     "valibot": "^1.1.0"
   },
   "devDependencies": {
